build 0.73.5 (READ DESC)
This build is to integrate the terminal system with the angular front end ON LOCAL MONGODB ONLY

This gets rid of the cluster connection and connects directly through localhost to a database on the operating machine.

Adding and removing data will now be reflected in the proper categories.

Further changes will be to update the integration